# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reliable Fresh Export Management System",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/lab-samples/queue": {
      "get": {
        "tags": [
          "lab sampling"
        ],
        "summary": "Lab Queue",
        "operationId": "lab_queue_api_v1_lab_samples_queue_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SeasonRegistrationRead"
                  },
                  "type": "array",
                  "title": "Response Lab Queue Api V1 Lab Samples Queue Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/registrations/{reg_id}/lab-sample": {
      "post": {
        "tags": [
          "lab sampling"
        ],
        "summary": "Record Lab Sample",
        "operationId": "record_lab_sample_api_v1_registrations__reg_id__lab_sample_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "reg_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Reg Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabSampleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabSampleRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "lab sampling"
        ],
        "summary": "Read Lab Sample",
        "operationId": "read_lab_sample_api_v1_registrations__reg_id__lab_sample_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "reg_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Reg Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabSampleRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lab-samples/{sample_id}/seal-photo": {
      "post": {
        "tags": [
          "lab sampling"
        ],
        "summary": "Upload Seal Photo",
        "operationId": "upload_seal_photo_api_v1_lab_samples__sample_id__seal_photo_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sample Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_seal_photo_api_v1_lab_samples__sample_id__seal_photo_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabSampleRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lab-samples/{sample_id}/documents": {
      "post": {
        "tags": [
          "lab sampling"
        ],
        "summary": "Upload Documents",
        "operationId": "upload_documents_api_v1_lab_samples__sample_id__documents_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sample Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_documents_api_v1_lab_samples__sample_id__documents_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabSampleRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/registrations/{reg_id}/harvests": {
      "post": {
        "tags": [
          "harvesting"
        ],
        "summary": "Record Harvest",
        "operationId": "record_harvest_api_v1_registrations__reg_id__harvests_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "reg_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Reg Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HarvestCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarvestRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "harvesting"
        ],
        "summary": "List Harvests",
        "operationId": "list_harvests_api_v1_registrations__reg_id__harvests_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "reg_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Reg Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HarvestRead"
                  },
                  "title": "Response List Harvests Api V1 Registrations  Reg Id  Harvests Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_documents_api_v1_lab_samples__sample_id__documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_documents_api_v1_lab_samples__sample_id__documents_post"
      },
...
      "Body_upload_seal_photo_api_v1_lab_samples__sample_id__seal_photo_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_seal_photo_api_v1_lab_samples__sample_id__seal_photo_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HarvestCreate": {
        "properties": {
          "harvest_date": {
            "type": "string",
            "format": "date",
            "title": "Harvest Date"
          },
          "supervisor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supervisor Name"
          },
          "supervisor_contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supervisor Contact"
          },
          "vehicle_trips": {
            "items": {
              "$ref": "#/components/schemas/VehicleTripCreate"
            },
            "type": "array",
            "minItems": 1,
            "title": "Vehicle Trips"
          }
        },
        "type": "object",
        "required": [
          "harvest_date",
          "vehicle_trips"
        ],
        "title": "HarvestCreate"
      },
      "HarvestRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "season_registration_id": {
            "type": "integer",
            "title": "Season Registration Id"
          },
          "harvest_date": {
            "type": "string",
            "format": "date",
            "title": "Harvest Date"
          },
          "supervisor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supervisor Name"
          },
          "supervisor_contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supervisor Contact"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "vehicle_trips": {
            "items": {
              "$ref": "#/components/schemas/VehicleTripRead"
            },
            "type": "array",
            "title": "Vehicle Trips",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "season_registration_id",
          "harvest_date",
          "supervisor_name",
          "supervisor_contact",
          "created_by",
          "created_at"
        ],
        "title": "HarvestRead"
      },
      "LabResult": {
        "type": "string",
        "enum": [
          "Pass",
          "Fail"
        ],
        "title": "LabResult"
      },
      "LabSampleCreate": {
        "properties": {
          "lab_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lab Name"
          },
          "sampling_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sampling Date"
          },
          "seal_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seal No"
          },
          "variety_confirmed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variety Confirmed"
          },
          "area_ha_2a": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Ha 2A"
          },
          "yield_4b_mt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yield 4B Mt"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          },
          "tss_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tss Value"
          },
          "result": {
            "$ref": "#/components/schemas/LabResult"
          }
        },
        "type": "object",
        "required": [
          "result"
        ],
        "title": "LabSampleCreate"
      },
      "LabSampleRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "season_registration_id": {
            "type": "integer",
            "title": "Season Registration Id"
          },
          "lab_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lab Name"
          },
          "sampling_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sampling Date"
          },
          "seal_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seal No"
          },
          "variety_confirmed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variety Confirmed"
          },
          "area_ha_2a": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Ha 2A"
          },
          "yield_4b_mt": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yield 4B Mt"
          },
          "seal_photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seal Photo Url"
          },
          "documents_2a4b_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documents 2A4B Url"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          },
          "tss_value": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tss Value"
          },
          "result": {
            "$ref": "#/components/schemas/LabResult"
          },
          "entered_by": {
            "type": "integer",
            "title": "Entered By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "season_registration_id",
          "lab_name",
          "sampling_date",
          "seal_no",
          "variety_confirmed",
          "area_ha_2a",
          "yield_4b_mt",
          "seal_photo_url",
          "documents_2a4b_url",
          "remark",
          "tss_value",
          "result",
          "entered_by",
          "created_at"
        ],
        "title": "LabSampleRead"
      },
...
      "RegistrationStatus": {
        "type": "string",
        "enum": [
          "Registered",
          "Field QC Passed",
          "Field QC Failed",
          "Lab Passed",
          "Lab Failed",
          "Under Contract",
          "Harvested (partial)",
          "Weighed",
          "Arrival QC Passed",
          "Arrival QC Failed",
          "Packed",
          "Palletised",
          "Pre-Cooled"
        ],
        "title": "RegistrationStatus",
        "description": "season_registrations.status \u2014 the central state machine.\nFull flow: PHASE_MAP.md Section 4.\n\nOnly covers Phases 1\u201311 (Registered through Pre-Cooled) since that's\neverything currently scoped. Finished Goods QC / Container Loaded /\nInvoiced / Export Documents Complete / Dispatched are explicitly\n(future) placeholders in PHASE_MAP.md \u2014 not added here until those\nphases are actually scoped (see PHASE_MAP.md Section 9, Open Question #13)."
      },
...
      "SeasonRegistrationRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "plot_id": {
            "type": "integer",
            "title": "Plot Id"
          },
          "season_year": {
            "type": "integer",
            "title": "Season Year"
          },
          "status": {
            "$ref": "#/components/schemas/RegistrationStatus"
          },
          "registered_by": {
            "type": "integer",
            "title": "Registered By"
          },
          "registered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Registered At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "plot_id",
          "season_year",
          "status",
          "registered_by",
          "registered_at"
        ],
        "title": "SeasonRegistrationRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VehicleTripCreate": {
        "properties": {
          "vehicle_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle No"
          },
          "driver_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Name"
          },
          "num_crates": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Crates"
          },
          "approx_weight_kg": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approx Weight Kg"
          }
        },
        "type": "object",
        "title": "VehicleTripCreate"
      },
      "VehicleTripRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "harvest_id": {
            "type": "integer",
            "title": "Harvest Id"
          },
          "vehicle_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle No"
          },
          "driver_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Name"
          },
          "num_crates": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Crates"
          },
          "approx_weight_kg": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approx Weight Kg"
          },
          "is_weighed": {
            "type": "boolean",
            "title": "Is Weighed",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "harvest_id",
          "vehicle_no",
          "driver_name",
          "num_crates",
          "approx_weight_kg"
        ],
        "title": "VehicleTripRead"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}